requirements-travis.txt
attrs<18
pyparsing
coverage
codecov
flake8
flake8-docstrings

